radvd: advertise a more sane route lifetime
Windows 10 drops an ipv6 route immediately when the advertised lifetime expires and it hasn't received a new advertisement.  After this happens a few times, the route gets dropped for good until the interface gets restarted.  With a 30 second lifetime, this was happening every few days.  With 300 I haven't seen the problem in over a week.